requirements.txt
# ── DataBridge ─────────────────────────────────────────
# PostgreSQL (OLTP) → DuckDB (OLAP) ETL Pipeline
# ───────────────────────────────────────────────────────

# Database drivers & ORM
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0

# Data processing
pandas>=2.1,<3.0
pyarrow>=14.0,<18.0

# Analytical data store
duckdb>=0.10,<2.0

# Dashboard
streamlit>=1.30,<2.0
plotly>=5.18,<6.0

# Excel parsing
openpyxl>=3.1,<4.0

# Config & environment
python-dotenv>=1.0,<2.0

# ── Dev / Testing ─────────────────────────────────────
pytest>=8.0,<9.0
pytest-cov>=4.1,<6.0
